Create packages
Create duke and exceptions packages. Exception classes are stored in package exceptions. The rest are stored in duke.
