+++ src/main/java/exceptions/FileUnloadableException.java
@@ -1,3 +1,5 @@
+package exceptions;
+
 public class FileUnloadableException extends DukeException {
     public FileUnloadableException(String msg) {
         super(msg);
