+++ src/main/java/exceptions/InvalidInputException.java
@@ -1,15 +1,14 @@
-public class OutOfRangeException extends DukeException {
-    String msg;
+package exceptions;
 
-    public OutOfRangeException(String msg) {
+public class InvalidInputException extends DukeException {
+    public InvalidInputException(String msg) {
         super(msg);
-        this.msg = msg;
     }
 
     @Override
     public String toString() {
         return "    ____________________________________________________________\n" +
-                "     ☹ OOPS!!! " + this.msg + " number out of range.\n" +
+                "     ☹ OOPS!!! I'm sorry, but I don't know what that means :-(\n" +
                 "    ____________________________________________________________";
     }
 }
