+++ src/main/java/exceptions/OutOfRangeException.java
@@ -1,12 +1,17 @@
-public class InvalidInputException extends DukeException {
-    public InvalidInputException(String msg) {
+package exceptions;
+
+public class OutOfRangeException extends DukeException {
+    String msg;
+
+    public OutOfRangeException(String msg) {
         super(msg);
+        this.msg = msg;
     }
 
     @Override
     public String toString() {
         return "    ____________________________________________________________\n" +
-                "     ☹ OOPS!!! I'm sorry, but I don't know what that means :-(\n" +
+                "     ☹ OOPS!!! " + this.msg + " number out of range.\n" +
                 "    ____________________________________________________________";
     }
 }
